+++ package.json
@@ -66,6 +66,7 @@
     "@testing-library/react": "13.4.0",
     "@types/react": "18.0.34",
     "@types/react-dom": "18.0.11",
+    "@types/uuid": "^9.0.2",
     "@typescript-eslint/eslint-plugin": "5.57.1",
     "@typescript-eslint/parser": "5.57.1",
     "@vitejs/plugin-react": "^4.0.0",
